Fix edges and data types

+++ include/operation/memcpy.h
@@ -17,7 +17,7 @@ struct Memcpy : public Operation {
 
   Memcpy() : Operation(0, 0, OPERATION_TYPE_MEMCPY) {}
 
-  Memcpy(u64 op_id, u32 ctx_id, uint32_t src_memory_op_id, uint32_t dst_memory_op_id,
+  Memcpy(u64 op_id, i32 ctx_id, u64 src_memory_op_id, u64 dst_memory_op_id,
          const std::string &hash, double redundancy, double overwrite)
       : Operation(op_id, ctx_id, OPERATION_TYPE_MEMCPY),
         src_memory_op_id(src_memory_op_id),
